The hybrid walk drive optimization. This implements the hybrid walk/drive implementation.
First EE motion while in contact are now parameterized and optimized over.
The number of polynoms are now split up in number of polynoms per stance and number of polynoms
per swing. They apply for both, force and motion of end effectors, but the force nodes are fixed at
zero during swing.

This implies changes to the terrain constraint, since position is not constant anymore during
contact. New is that the scalar product of the velocity node with the terrain normal is constraint
to zero fore all contact nodes. This assumes, that the terrain between two consecutive contact nodes
can be approximated by a polynom. Tests with the current terrain examples seem to suggest, that
that is a reasonamble assumtion to make.

Renaming of the variable constant to contact was made to avoid confusion and for better code
readability. Note during contact end effector motion nodes are not necessarly constant anymore.